% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdbk.R
\name{cdbk_df}
\alias{cdbk_df}
\alias{cdbk_dt}
\title{Codebook}
\usage{
cdbk_df(data)

cdbk_dt(data, ...)
}
\arguments{
\item{data}{labelled tibble}

\item{...}{other arguments passed to \code{\link[DT:datatable]{DT::datatable()}}}
}
\value{
Function \code{\link[=cdbk_df]{cdbk_df()}} returns a tibble with columns:
\itemize{
\item \code{variable} - variable name
\item \code{label} - variable label
\item \code{value_labels} - list of tibbles with value labels
}

Function \code{\link[=cdbk_dt]{cdbk_dt()}} uses \code{\link[DT:datatable]{DT::datatable()}} to generate a browsable
and searchable codebook based on the documentation.
}
\description{
Generate a codebook based on data documentation (variable names, variable
labels, and value labels).
}
