% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq.R
\name{freq_df}
\alias{freq_df}
\title{Frequency tibble}
\usage{
freq_df(data, var = -1, name = NULL, ...)
}
\arguments{
\item{data}{data frame}

\item{var}{Variable name or index (negative counting from the right, positive
counting from the left), see \code{\link[dplyr:pull]{dplyr::pull()}}}

\item{name}{Column which values will be used as names. Specified as \code{var}. C.f. \code{\link[dplyr:pull]{dplyr::pull()}}}

\item{...}{other arguments}
}
\description{
Frequency tibble
}
