% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misvals.R
\name{misvals}
\alias{misvals}
\alias{misranges}
\alias{misranges_df}
\title{Lists and tibbles with user missing values}
\usage{
misvals(data)

misranges(data)

misranges_df(data)
}
\arguments{
\item{data}{Labelled tibble with missing values/ranges defined}
}
\value{
Function \code{\link[=misvals]{misvals()}} returns a list of vectors of missing values (or
\code{NULL}s if none are defined). List elements are named with variable names.

Function \code{\link[=misranges]{misranges()}} returns a list of two-element vectors with
ranges of missing values.

Function \code{\link[=misranges_df]{misranges_df()}} returns a tibble with columns
\itemize{
\item \code{variable} - Variable name
\item \code{from}, \code{to} - Bounds (inclusive) of the missing value range
}
}
\description{
Lists and tibbles with user missing values
}
