% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seems.R
\name{seems}
\alias{seems}
\alias{seems_integer}
\alias{seems_continuous}
\alias{seems_discrete}
\alias{seems_categorical}
\title{Heuristically guess or verify type of a variable}
\usage{
seems_integer(x)

seems_continuous(x)

seems_discrete(x)

seems_categorical(x)
}
\arguments{
\item{x}{a vector for which \code{\link[=is.atomic]{is.atomic()}} is \code{TRUE}}
}
\value{
All \verb{seems_*()} functions return \code{TRUE} or \code{FALSE} (a logical
scalar).
}
\description{
Given a vector \code{x} guess what kind of "type" it is, where "type"
correspond to some common classes influencing useful ways to
analyze/visualize. These functions are used internally to determine default
behavior of some other functions.

\itemize{
\item \code{seems_integer()} -- Returns \code{TRUE} if \code{x} is \code{\link[=typeof]{typeof()}}
integer or is a numeric with all values being in fact integers (i.e. equal
to \code{round(x)}).
}

\itemize{
\item \code{seems_continuous()} -- Returns \code{TRUE} if \code{x} is numeric and,
if an integer, has more than 10 distinct values.
}

\itemize{
\item \code{seems_discrete()} -- Returns \code{TRUE} if \code{x} does not seem to
be continuous.
}

\itemize{
\item \code{seems_categorical()} -
}
}
\examples{
# seems_integer() ----------------------------------------------------------
seems_integer(1:5)       # TRUE
seems_integer(runif(5))  # FALSE

# seems_continuous() -------------------------------------------------------
seems_continuous(1:5)       # FALSE
seems_continuous(1:11)      # TRUE
seems_continuous(runif(5))  # TRUE

# Summarize variables of `mtcars` in a type-dependent way by drawing a
# histogram for continuous ones and barchart with counts for non-continuous
# ones.
iscont <- vapply(mtcars, seems_continuous, logical(1))
layout(matrix(1:12, 3, 4))
for(n in names(mtcars)) {
  if(iscont[n]) {
    hist(mtcars[[n]], main=n, xlab="")
  } else {
    barplot(table(mtcars[[n]]), main=n)
  }
}
layout(1)
}
