% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varlabs.R
\name{varlabs}
\alias{varlabs}
\title{Data frame of variable labels}
\usage{
varlabs(data)
}
\arguments{
\item{data}{data frame with labelled variables}
}
\value{
A tibble with columns
\itemize{
\item \code{variable} - Variable name
\item \code{label} - Variable label
}
}
\description{
Data frame of variable labels
}
\seealso{
\code{\link[labelled:var_label]{labelled::var_label()}}
}
