% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vallabs.R
\name{vallabs}
\alias{vallabs}
\title{List of tibbles with value labels}
\usage{
vallabs(data)
}
\arguments{
\item{data}{data frame with variables having value labels}
}
\value{
A list of tibbles (or \code{NULL}s of no value labels are present) named
with variable names. Each tibble has columns:
\itemize{
\item \code{label} - Value label
\item \code{value} - Value
}
}
\description{
List of tibbles with value labels
}
\seealso{
\code{\link[labelled:val_labels]{labelled::val_labels()}}
}
